requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
gigachat==0.1.8